requirements.txt
jupyter
jupyterlab
numpy
matplotlib
tweepy
python-dotenv
pymongo
dnspython
click
